Subsettings component was moved to Settings namespace
