--- src/layouts/SubsettingsLayout/index.js
@@ -1,5 +1,0 @@
-// @flow
-
-import SubsettingsLayout from './SubsettingsLayoutContainer'
-
-export default SubsettingsLayout
