+++ src/layouts/index.js
@@ -3,11 +3,9 @@
 import CoreLayout from './CoreLayout'
 import MenuLayout from './MenuLayout'
 import WalletsLayout from './WalletsLayout'
-import SubsettingsLayout from './SubsettingsLayout'
 
 export {
   CoreLayout,
   MenuLayout,
   WalletsLayout,
-  SubsettingsLayout,
 }
