+++ src/routes/Settings/components/SubsettingsView/SubsettingsViewContainer.js
@@ -3,11 +3,11 @@ import { connect } from 'react-redux'
 
 import reactRouterBack from 'utils/browser/reactRouterBack'
 
-import SubsettingsLayout from './SubsettingsLayout'
+import SubsettingsView from './SubsettingsView'
 
 const mapDispatchToProps = {
   closeClick: () => reactRouterBack({ fallbackUrl: '/settings' }),
 }
 export default (
-  connect/* :: < AppState, any, Object, _, _ > */(null, mapDispatchToProps)(SubsettingsLayout)
+  connect/* :: < AppState, any, Object, _, _ > */(null, mapDispatchToProps)(SubsettingsView)
 )
