+++ src/routes/Settings/components/SubsettingsView/index.js
@@ -1,0 +1,5 @@
+// @flow
+
+import SubsettingsView from './SubsettingsViewContainer'
+
+export default SubsettingsView
