+++ src/routes/Settings/components/SubsettingsView/subsettingsView.scss
@@ -1,4 +1,5 @@
 .subsettings-layout {
+  all: initial;
   position: relative;
   height: 100%;
   display: flex;
