+++ src/routes/Settings/components/index.js
@@ -1,0 +1,7 @@
+// @flow
+
+import SubsettingsView from './SubsettingsView'
+
+export {
+  SubsettingsView,
+}
