+++ src/routes/Settings/routes/SecurityPassword/SecurityPassword.js
@@ -2,14 +2,12 @@
 
 import React from 'react'
 
-import { SubsettingsLayout } from 'layouts'
+import { SubsettingsView } from 'routes/Settings/components'
 
 export default function SecurityPassword() {
   return (
-    <div className='settings-view'>
-      <SubsettingsLayout title='Update security password'>
-        <h1>Hello world</h1>
-      </SubsettingsLayout>
-    </div>
+    <SubsettingsView title='Update security password'>
+      <h1>Hello world</h1>
+    </SubsettingsView>
   )
 }
