+++ src/styles/core.scss
@@ -142,12 +142,12 @@
   @import '../routes/NotFound/notFoundView';
   @import '../routes/DigitalAssets/digitalAssets';
   @import '../routes/Transactions/transactionsView';
+  @import '../routes/Settings/components/SubsettingsView/subsettingsView';
 
   // Layouts
   @import '../layouts/CoreLayout/coreLayout';
   @import '../layouts/MenuLayout/menuLayout';
   @import '../layouts/WalletsLayout/walletsLayout';
-  @import '../layouts/SubsettingsLayout/subsettingsLayout';
 
   // To be removed
   // @import '../layouts/ActiveAssetsLayout/activeAssetsLayout';
